🗺️ Atlas: [architectural change]
🕸️ Tangle: `redaction` imported from `stream` (`StreamEvent`, `StreamSink`) to provide `RedactingSink`, while `stream` heavily depended on `redaction` for the core redactor, forming a cycle.
📐 Blueprint: Extracted `RedactingSink` out of `src/redaction.rs` into its own module `src/stream/redacting.rs`. Now, `stream` depends on `redaction`, but `redaction` no longer needs `stream`.
🧱 Stability: A cleaner dependency graph that adheres to strict unidirectional flows. Also fixes several unwrap statements and cargo doc dead-links as standard maintenance.
🔬 Verification: Confirmed graph acyclicity manually via test scripts. `cargo fmt --all && cargo clippy --all-targets --all-features -- -D warnings && cargo test --lib && cargo doc --no-deps` passed seamlessly.
